chore: Clean up `experiments/ezpz/train.py`
