+++ torchtitan/experiments/ezpz/logging.py
@@ -44,4 +44,3 @@ def warn_once(logger: logging.Logger, msg: str) -> None:
     if msg not in _logged:
         logger.warning(msg)
         _logged.add(msg)
-
